turned a static create ReadPileup method into a constructor
