+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/annotator/VariantAnnotator.java
@@ -237,7 +237,7 @@ public class VariantAnnotator extends VariantWalker {
         final AlleleLikelihoods<GATKRead, Allele> result = new AlleleLikelihoods<>(variantSamples,
                 new IndexedAlleleList<>(vc.getAlleles()), AssemblyBasedCallerUtils.splitReadsBySample(variantSamples, getHeaderForReads(), reads));
 
-        final ReadPileup readPileup = GATKProtectedVariantContextUtils.getPileup(vc, readsContext);
+        final ReadPileup readPileup = new ReadPileup(vc, readsContext);
         final Map<String, ReadPileup> pileupsBySample = readPileup.splitBySample(getHeaderForReads(), "__UNKNOWN__");
 
         final Allele refAllele = vc.getReference();
