+++ src/test/java/org/broadinstitute/hellbender/utils/pileup/ReadPileupUnitTest.java
@@ -675,4 +675,43 @@ public final class ReadPileupUnitTest extends GATKBaseTest {
         Assert.assertFalse(filteredReadsKeepDiscordant.contains(read1BasesAgree), "The lower quality base was kept.");
         Assert.assertTrue(filteredReadsKeepDiscordant.contains(read2BasesAgree), "The higher quality base is missing.");
     }
+
+
+    @Test
+    public void testGetPileup() {
+        final SAMFileHeader header = ArtificialReadUtils.createArtificialSamHeader(1, 1, 1000);
+        final Locatable loc = new SimpleInterval("chr1", 10, 10);
+        final int readLength = 3;
+
+        //this read doesn't overlap {@code loc}
+        final GATKRead read1 = ArtificialReadUtils.createArtificialRead(header, "read1", 0, 1, readLength);
+        read1.setBases(Utils.dupBytes((byte) 'A', readLength));
+        read1.setBaseQualities(Utils.dupBytes((byte) 30, readLength));
+        read1.setCigar("3M");
+
+        //this read overlaps {@code loc} with a Q30 'A'
+        final GATKRead read2 = ArtificialReadUtils.createArtificialRead(header, "read2", 0, 10, readLength);
+        read2.setBases(Utils.dupBytes((byte) 'A', readLength));
+        read2.setBaseQualities(Utils.dupBytes((byte) 30, readLength));
+        read2.setCigar("3M");
+
+        //this read overlaps {@code loc} with a Q20 'C' (due to the deletion)
+        final GATKRead read3 = ArtificialReadUtils.createArtificialRead(header, "read3", 0, 7, readLength);
+        read3.setBases(Utils.dupBytes((byte) 'C', readLength));
+        read3.setBaseQualities(Utils.dupBytes((byte) 20, readLength));
+        read3.setCigar("1M1D2M");
+
+        //this read doesn't overlap {@code loc} due to the deletion
+        final GATKRead read4 = ArtificialReadUtils.createArtificialRead(header, "read4", 0, 7, readLength);
+        read4.setBases(Utils.dupBytes((byte) 'C', readLength));
+        read4.setBaseQualities(Utils.dupBytes((byte) 20, readLength));
+        read4.setCigar("1M5D2M");
+
+        final ReadPileup pileup = new ReadPileup(loc, Arrays.asList(read1, read2, read3, read4));
+
+        // the pileup should contain a Q30 'A' and a Q20 'C'
+        final int[] counts = pileup.getBaseCounts();
+        Assert.assertEquals(counts, new int[]{1, 1, 0, 0});
+
+    }
 }
